+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Full Change Log
 
+- [v0.8.2](https://github.com/Windos/BurntToast/releases/download/v0.8.2/BurntToast.zip)
+
+  - Add: AdaptiveGroups are now usable via New-BTColumn
+
 - [v0.8.1](https://github.com/Windos/BurntToast/releases/download/v0.8.1/BurntToast.zip)
 
   - Fix: Toast alias removed in 0.8.0 has been restored
@@ -8,7 +12,7 @@
 
   - Deprecation: Signalling removal of Path parameter from New-BTAudio in future version, v0.9.0.
 
-    - https://github.com/MicrosoftDocs/windows-uwp/issues/1593
+    - See [MicrosoftDocs/windows-uwp Issue #1593](https://github.com/MicrosoftDocs/windows-uwp/issues/1593)
 
 - [v0.8.0](https://github.com/Windos/BurntToast/releases/download/v0.8.0/BurntToast.zip)